requirements.txt
Django==2.1.1
django-filebrowser-no-grappelli==3.7.5
django-jet==1.0.7
django-taggit==0.23.0
django-tinymce4-lite==1.7.2
jsmin==2.2.2
Pillow==5.2.0
pytz==2018.5
sorl-thumbnail==12.5.0
